Adding css styles to all files
