--- tweet-client/src/components/NavLink/navlink.css
@@ -1,61 +1,0 @@
- .Navlink{
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 50px;
-    padding: 10px 10px;
-    background-color: teal;
-    color:white;
-    font-family: Raleway;
-    width:100%;
-}
-
-.item{
-    display: flex;
-}
-.item li{
-    margin-right: 30px;
-    list-style: none;
-    display: block;
-    height: 100%;
-    padding: 5px;
-    width: 100%;
-}
-.item li a{
-    text-decoration: none;
-    color:white;
-    box-sizing: border-box;
-    display: block;
-}
-item li a:active{
-    background-color:#006666;
-    height: 100%;
-    width:100%;
-    border-bottom: 3px solid orange;
-}
-
-.item li:hover{
-    background-color:  	#004c4c;
-    border-bottom: 3px solid orange;
-}
-.brand{
-    padding:10px 10px;
-    text-decoration: none;
-    list-style: none;
-    margin-left: 20px;
-}
-
-.brand:hover{
-    background-color: #003434;
-    border-bottom: 3px solid orange;
-}
-
-.brand li a{
-    text-decoration: none;
-    color: whitesmoke;
-}
-
-
-
-@media(max-width:500){
-
-}
