--- tweet-client/src/components/auth.css
@@ -1,46 +1,0 @@
-
-.container{
-    text-align: center;
-}
-
-.input{
-    margin: 10px 10px;
-    outline: none;
-    padding: 10px;
-}
-.form{
-    padding: 10px;
-    width: 500px;
-    height: 600px;
-    margin-top: 100px;
-    margin-bottom: 20px;
-    margin:auto auto;
-    display: block;
-    background-color:#eee;
-    border: 3px solid #ccc;
-}
-
-.form h1{
-    color: #111E6C;
-}
-
-.submit{
-    width: 40%;
-    margin: 10px;
-    height: 30px;
-    color: #111E6C;
-    font-size: 16px;
-}
-
-.submit:hover{
-    background-color:#008CBA;
-    color: whitesmoke;
-    border-bottom:3px solid orange;
-}
-
-@media(max-width:500px){
-    .form{
-        width:300px;
-        height: 100vh;
-    }
-}
