--- tweet-client/src/components/message/add.message.css
@@ -1,35 +1,0 @@
-.form{
-    margin-left: 33%;
-}
-.item {
-    margin:20px 10px;
-    font-size: 30px;
-    color: royalblue;
-}
-form{
-    border: 2px solid #ccc;
-    background-color: #eee;
-    width: 500px;
-    text-align: center;
-    font-size: 30px;
-    height: 300px;
-    margin-top: 10%;
-}
-textarea{
-    margin-left: 0px;
-}
-
-.submit{
-    width: 40%;
-    margin: 20px;
-    height: 30px;
-    color: #111E6C;
-    font-size: 16px;
-}
-
-.submit:hover{
-    background-color:#008CBA;
-    color: whitesmoke;
-    border-bottom:3px solid orange;
-}
-
