+++ tweet-client/src/container/Login/LoginPage.js
@@ -1,8 +1,9 @@
 import React,{Component} from 'react'
 import {connect} from 'react-redux'
-import  { signin } from '../action'
+import  { signin } from '../../action'
 import classes from './auth.css'
 import {Redirect} from 'react-router-dom'
+import {Link} from 'react-router-dom'
 
 class Login extends Component{
     constructor(props){
@@ -32,7 +33,7 @@ class Login extends Component{
         return (
             <div className={classes.container}>
             {redirect}
-                <form className={classes.form} onSubmit={(event)=>this.onsubmithandler(event)}>
+                {/* <form className={classes.form} onSubmit={(event)=>this.onsubmithandler(event)}>
                 <h1>Login</h1>
                         <div>Email:</div><input className={classes.input} type="text" name='email' onChange={(event)=>this.onchange(event)}/>
                         <div>UserName:</div><input className={classes.input} type="text" name="username" onChange={(event)=> this.onchange(event)} placeholder="Enter your UserName" />    
@@ -41,6 +42,25 @@ class Login extends Component{
                     <input className={classes.submit} type="submit" value="Login"/>
 
                 </div>
+                </form> */}
+                <form className={classes.form} onSubmit={(event)=>this.onsubmithandler(event)}>
+                    <div className={classes.item}>
+                        <div>
+                            <input className={classes.input} type="text" name='email' onChange={(event)=>this.onchange(event)}  placeholder="Enter your email"/>
+
+                        </div>
+                        <div>
+                            <input className={classes.input} type="text" name="username" onChange={(event)=> this.onchange(event)} placeholder="Enter your UserName" />
+
+                        </div>
+                        <div>
+                            <input className={classes.input} type="password"  name='password' onChange={(event)=> this.onchange(event)} placeholder="Enter your Password" /> 
+                        </div>
+                        <div>
+                            <input className={classes.submit} type="submit" value="Login"/>
+                        </div>
+                        <span className = {classes.signup}>Not Registered</span><span className = {classes.signup} ><Link  to="/signup">signup</Link></span> 
+                     </div>
                 </form>
             </div>
         )
