+++ tweet-client/src/container/Signup/auth.css
@@ -1,0 +1,74 @@
+.container{
+    height: 100vh;
+    width: 100%;
+    border: 1px solid #008060;
+    /* background: url('../../assests/techno5-0.png');
+    background-repeat: no-repeat;
+    background-size: 100% 100%; */
+    background-color:#008060;
+}
+
+.form{
+    margin:auto ;
+    margin-top: 0%; 
+    background-color: white;
+    box-sizing: border-box;
+    box-shadow:    2px 3px 2px  #ccc;
+    animation: move 2s ease-out  1 forwards; 
+}
+.item{
+    margin-top: 20px;
+}
+
+.input{
+ 
+        width: 50%;
+        padding: 5px;
+        margin: 8px 0;
+        box-sizing: border-box;
+        border: 1px solid gray;
+    
+    
+}
+
+.submit{
+    margin-top: 20px;
+    width: 55%;
+    padding: 10px;
+    font-size: 20px;
+    background-color: green;
+    color: white;
+    outline: none;
+    border: none;
+    cursor: pointer;
+    border-radius: 20px;
+    box-shadow: 0px 3px green;
+    position: relative;
+}
+
+.signup{
+    font-size: 15px;
+    color:gray;
+    margin-right: 5px;
+}
+
+.submit:hover{
+    background-color: darkgreen;
+    box-shadow: 0px 2px green;
+    top:2px;
+}
+
+@keyframes move{
+    0%{
+        background-color: black;
+        opacity: 0;
+    }
+    50%{
+        opacity: 0.5;
+        background-color: gray;
+    }
+    100%{
+        transform: translateY(50px);
+        opacity: 1;
+    }
+}
