+++ tweet-client/src/container/messages/messages.css
@@ -1,14 +1,3 @@
-.img{
-    height: 100%;
-    width: 100%;
-}
-
-.usermessage{
-    display: flex;
-    border:1px solid #ccc;
-    font-family: Raleway;
-    box-shadow: 2px 3px #eee;
-}
 
 .item{
     margin-left: 20px;
@@ -21,5 +10,14 @@
     }
     .usermessage{
         width: 250px;
+    }
+}
+
+@keyframes move{
+    0%{
+        transform: translateX(400px);
+    }
+    100%{
+        transform: translateX(70px);
     }
 }
